First pass at core_file creation from the information we receive from Drupal.  Note that the multiple saves of each object are actually a system requirement - the RELS-EXT datastream fails to populate correctly with non-persisted objects that have rels defined on them.

+++ app/services/core_file_creator_service.rb
@@ -1,30 +1,61 @@
 class CoreFileCreatorService
   include Concerns::TapasObjectCreator
 
-  def create_record 
+  def create_record(run_cleanup = true) 
     # Things to do:
       # 1. Instantiate record with pid assigned
       # 2. Assign metadata from all params
       # 3. If a TEI-XML object type is provided, extract the TEI header
       # 4. Assign all file objects to their own content_file objects
       # 5. Return a 201 to some endpoint on the drupal site. 
+    begin 
+      core     = nil 
+      tei_file = nil
 
-    core     = nil 
-    tei_file = nil
+      # 1. Instantiate a core record 
+      core = CoreFile.new
 
-    # 1. Instantiate a core record 
-    core = CoreFile.new
+      # 2. Assign metadata from all params
+      core.depositor        = params[:depositor]
+      core.mass_permissions = "private" 
+      core.nid              = params[:node_id]
+     
+      core.save!
 
-    # 2. Assign metadata from all params
-    core.depositor  = params[:depositor]
-    # TODO: How do collections work now?
+      # Attach this file to its collection, or add it to the phantom 
+      # collection if it doesn't seem to have one.
+      collection = Collection.find_by_nid(params[:collection_id])
+      if collection
+        core.collection_id = collection.id
+      else
+        core.collection_id = Collection.phantom_collection.pid
+      end
 
-    # Take params[:file] (always assumed to be a string file path as 
-    # this makes enqueing this in a job possible), extract it to its 
-    # own TEIFile object, and kick off inline derivation creation
-    tei_file = TEIFile.new(depositor: params[:depositor])
-    file_str = File.read(params[:file])
-    tei_file.content = file_str
-    FileUtils.rm(params[:file])
+      core.save!
+
+      # Take params[:file] (always assumed to be a string file path as 
+      # this makes enqueing this in a job possible), extract it to its 
+      # own TEIFile object, and kick off inline derivation creation
+      tei_file = TEIFile.new(depositor: params[:depositor])
+
+      fname = Pathname.new(params[:file]).basename.to_s
+      fblob = File.read(params[:file])
+      tei_file.add_file(fblob, "content", fname)
+      tei_file.depositor = params[:depositor]
+      tei_file.canonize
+      tei_file.save!
+      tei_file.core_file = core
+      tei_file.save!
+      return core
+    rescue => e
+      # If the process doesn't complete, ensure files are cleaned out of the 
+      # repository and that an email is sent (production environments only).
+      core.destroy if core && core.persisted? 
+      tei_file.destroy if tei_file && tei_file.persisted?
+      raise e
+    ensure
+      # Always ensure that the file at params[:file] is deleted.
+      FileUtils.rm(params[:file]) if run_cleanup
+    end
   end
 end
